Iron Man fixes and Prefab fixes
Fixed some of the issues regarding the Iron Man level, i.e. the color of the activators, along with adding the columns for note color correction. Also fixed some issues that existed with the prefabs, such as the columns not having "isTrigger" checked, along with the activator prefabs themselves lacking the same trigger setting.